Extracted Hand class to separate file.
